backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pydantic
pytesseract
pdf2image
Pillow

